pyproject.toml

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automationaccelerator"
version = "0.1.0"
description = "Mainframe modernization accelerators: utility scanners, conversion log reports, IMS analyzer."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Rajesh Kumar"}]
dependencies = [
  "pandas>=2.0",
  "pyyaml>=6",
  "openpyxl>=3.1",
  "xlsxwriter>=3.0"
]

[project.scripts]
autoaccel = "automationaccelerator.__main__:main"  # entrypoint: runs __main__ module

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
automationaccelerator = [
  "config/*.yaml",
  "config/*.txt",
  "data/*",
  "*.yaml",
  "*.yaml.bak"
]
